refactor: svg 아이콘 파일 및 index.ts 파일을 svg 폴더 하위로 이동
