+++ frontend/src/assets/svg/index.ts
